Compile tracker separetly from website assets
