+++ .gitignore
@@ -27,6 +27,7 @@ npm-debug.log
 
 # The directory NPM downloads your dependencies sources to.
 /assets/node_modules/
+/tracker/node_modules/
 
 # Since we are building assets from assets/,
 # we ignore priv/static. You may want to comment
